Fix duplicate stat records from Interrogator
If a team has fewer than 46 active players, FOF was clearing the player
index, but leaving the player id alone from the prior game. This would
cause a player to get 2 game entries created. Detect this case and clear
the player ID, and use this to not write out a stat record for such
players.